+++ tests/PSDocs.Tests/PSDocs.Common.Tests.ps1
@@ -163,3 +163,40 @@ Describe 'Invoke-PSDocument' -Tag 'FromPath' {
         }
     }
 }
+
+Describe 'New-PSDocumentOption' -Tag 'Option' {
+
+    Context 'Read PSDocs.yml' {
+
+        It 'can read default YAML' {
+            $option = New-PSDocumentOption;
+            $option.Generator | Should -Be 'PSDocs';
+        }
+    }
+
+    Context 'Read Markdown.Encoding' {
+
+        It 'from Hashtable' {
+            $option = New-PSDocumentOption -Option @{ 'Markdown.Encoding' = 'UTF8' };
+            $option.Markdown.Encoding | Should -Be 'UTF8';
+        }
+
+        It 'from YAML' {
+            $option = New-PSDocumentOption -Option (Join-Path -Path $here -ChildPath 'PSDocs.Tests.yml');
+            $option.Markdown.Encoding | Should -Be 'UTF8';
+        }
+    }
+
+    Context 'Read Markdown.WrapSeparator' {
+
+        It 'from Hashtable' {
+            $option = New-PSDocumentOption -Option @{ 'Markdown.WrapSeparator' = 'ZZZ' };
+            $option.Markdown.WrapSeparator | Should -Be 'ZZZ';
+        }
+
+        It 'from YAML' {
+            $option = New-PSDocumentOption -Option (Join-Path -Path $here -ChildPath 'PSDocs.Tests.yml');
+            $option.Markdown.WrapSeparator | Should -Be 'ZZZ';
+        }
+    }
+}